fix(booking-conflicts): fix booking conflict not showing on kit
